worker_service/requirements_worker.txt
Flask==2.2.3
langchain==0.0.309
openai==0.27.7
pydub==0.25.1
transformers==4.31.0
werkzeug==2.2.3
ffmpeg==1.4
flask-cors==4.0.0
google-cloud-storage
google-cloud-pubsub
google-cloud-tasks
ffprobe
ctransformers
torch